mds: abstract SessionMapStore from SessionMap
This is similar to what I did for InodeStore a while back:
introduce a logical separation between the persisted attributers
(and their encoding) and the live/runtime behavioural code.  This
results in a handy SessionMapStore class that can be used for
encode/decode from tools.

Also give it a reset_state method so that it matches the
prototype of the MDSTable subclasses for the benefit of
cephfs-table-tool.

Signed-off-by: John Spray <john.spray@redhat.com>